pyproject.toml
[tool.black]
line-length = 200

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 200

[tool.uv.sources]
django-des = { git = "https://github.com/jamiecounsell/django-des", rev = "4d121c62ffd0cd7e6a7288a6f6aa14e82fb247e1" }


[project]
name = "ordersmanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.10.0",
    "apscheduler==3.11.0",
    "asgiref==3.9.1",
    "attrs==25.3.0",
    "autobahn==24.4.2",
    "automat==25.4.16",
    "cffi==1.17.1",
    "cfgv==3.4.0",
    "channels==4.2.2",
    "click==8.2.1",
    "constantly==23.10.4",
    "cryptography==45.0.6",
    "daphne==4.1.2",
    "distlib==0.4.0",
    "django==4.2.20",
    "django-channels-presence-4-0==1.1.2",
    "django-des",
    "django-invitations>=2.1.0",
    "django-solo==2.3.0",
    "filelock==3.19.1",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httptools==0.6.4",
    "hyperlink==21.0.0",
    "identify==2.6.13",
    "idna==3.10",
    "incremental==24.7.2",
    "nodeenv==1.9.1",
    "packaging==25.0",
    "platformdirs==4.3.8",
    "pre-commit==4.3.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pyopenssl==25.1.0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "service-identity==24.2.0",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "twisted==25.5.0",
    "txaio==25.6.1",
    "typing-extensions==4.15.0",
    "tzlocal==5.3.1",
    "uvicorn==0.34.2",
    "uvicorn-worker==0.3.0",
    "uvloop==0.21.0",
    "virtualenv==20.34.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
    "zope-interface==7.2",
]
